+++ docs/changelog.rst
@@ -18,6 +18,7 @@ Unreleased
 - Use "import as" to prioritize the modules for editors.
 - Added parameter ``encoder_cls`` for ``jwt.encode`` and ``decoder_cls`` for ``jwt.decode``.
 - Added ``generate_new_key`` and ``remove_keys`` methods on ``jwk.KeySet``
+- Added additional ``jrfc_created_at`` and ``jrfc_disabled`` fields to ``KeyParameters`` to aid key rotation
 
 **Breaking changes**: